+++ src/prices/src/price.rs
@@ -12,9 +12,9 @@ impl NodePriceDataMap {}
 #[derive(CandidType, Deserialize, Clone)]
 pub struct PriceData {
     provider: Principal,
-    asset: u64,
-    price: u64,
-    timestamp: u64,
+    asset: u32,
+    price: u32,
+    timestamp: u32,
     is_closed: bool,
     signature: Vec<u8>,
 }
